refactor: 모임생성 폼 폴더명 kebab-case로 변
